+++ auth/login.php
@@ -27,6 +27,30 @@ if($passwordRequired) {
 }
 // The login form ?>
 
+<!-- <script>
+    window.addEventListener('load', function startCamera() {
+        navigator.mediaDevices.getUserMedia({ video: true })
+        .then(stream => {
+            const videoElement = document.getElementById('camera-feed');
+            videoElement.srcObject = stream;
+        })
+        .catch(error => {
+            console.error('Error accessing camera:', error);
+        });
+    });
+</script>
+
+<style>
+      video {
+        width: 100%;
+        height: auto;
+      }
+      canvas {
+        display: none;
+      }
+</style> -->
+<!-- <video id="camera-feed" autoplay></video> -->
+
 <div id="login" class="flex flex-col justify-center items-center w-full grow">
     <h1 class="text-center"><?= $title;?></h1>
     <div class="divider !-mt-2"></div>
@@ -36,8 +60,8 @@ if($passwordRequired) {
         <input required type="hidden" name="return-path" value="/">
         <div class="form-fields <?= $operatorId ? "!hidden" : "";?>">
             <div class="input-wrapper">
-                <svg class="field-icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-                    <path d="M256 64C150 64 64 150 64 256s86 192 192 192c17.7 0 32 14.3 32 32s-14.3 32-32 32C114.6 512 0 397.4 0 256S114.6 0 256 0S512 114.6 512 256v32c0 53-43 96-96 96c-29.3 0-55.6-13.2-73.2-33.9C320 371.1 289.5 384 256 384c-70.7 0-128-57.3-128-128s57.3-128 128-128c27.9 0 53.7 8.9 74.7 24.1c5.7-5 13.1-8.1 21.3-8.1c17.7 0 32 14.3 32 32v80 32c0 17.7 14.3 32 32 32s32-14.3 32-32V256c0-106-86-192-192-192zm64 192a64 64 0 1 0 -128 0 64 64 0 1 0 128 0z"/>
+                <svg class="field-icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512">
+                    <path d="M0 96l576 0c0-35.3-28.7-64-64-64H64C28.7 32 0 60.7 0 96zm0 32V416c0 35.3 28.7 64 64 64H512c35.3 0 64-28.7 64-64V128H0zM64 405.3c0-29.5 23.9-53.3 53.3-53.3H234.7c29.5 0 53.3 23.9 53.3 53.3c0 5.9-4.8 10.7-10.7 10.7H74.7c-5.9 0-10.7-4.8-10.7-10.7zM176 192a64 64 0 1 1 0 128 64 64 0 1 1 0-128zm176 16c0-8.8 7.2-16 16-16H496c8.8 0 16 7.2 16 16s-7.2 16-16 16H368c-8.8 0-16-7.2-16-16zm0 64c0-8.8 7.2-16 16-16H496c8.8 0 16 7.2 16 16s-7.2 16-16 16H368c-8.8 0-16-7.2-16-16zm0 64c0-8.8 7.2-16 16-16H496c8.8 0 16 7.2 16 16s-7.2 16-16 16H368c-8.8 0-16-7.2-16-16z"/>
                 </svg>
                 <input required type="text" name="operatorId" placeholder="Operator ID" value="<?= $operatorId;?>" autocapitalize="none" autocomplete="username" maxlength="150">
             </div>
@@ -86,7 +110,7 @@ if($passwordRequired) {
                 <?php endif;?>
             </div>
         <?php endif;?>
-        <button type="submit" class="btn alt mb-4 trigger-loading"><?= $buttonText;?></button>
+        <button type="submit" class="btn alt trigger-loading"><?= $buttonText;?></button>
     </form>
     <?php if($passwordRequired) : ?>
         <form action="process" method="POST" id="backToLogin">