+++ views/about.php
@@ -1,8 +1,13 @@
-<h1>About Cain</h1>
+<h1>About DRW</h1>
 
-<section class="notice">
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-        <path d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM216 336h24V272H216c-13.3 0-24-10.7-24-24s10.7-24 24-24h48c13.3 0 24 10.7 24 24v88h8c13.3 0 24 10.7 24 24s-10.7 24-24 24H216c-13.3 0-24-10.7-24-24s10.7-24 24-24zm40-208a32 32 0 1 1 0 64 32 32 0 1 1 0-64z"/>
-    </svg>
-    <p>Need to write some stuff about Cain here.</p>
-</section>
+<p>DRW (Europe) Ltd, Science Village, Chesterford Research Park, Little Chesterford, CB10 1XL, UK</p>
+<p>Tel: <a class="alt" href="tel:+441223784332">+44 1223 784332</a></p>
+<p>Email: <a class="alt" href="mailto:sales@drw-ltd.com">sales@drw-ltd.com</a></p>
+<br>
+<p>DRW (US) Ltd, 845 Embedded Way, Suite 150, San Jose, CA 95138, USA</p>
+<p>Tel: <a class="alt" href="tel:+14087731511">+1 408-773-1511</a></p>
+<p>Fax: +1 408-773-1553</p>
+<p>Email: <a class="alt" href="mailto:sales@drw-ltd.com">sales@drw-ltd.com</a></p>
+<p>EIN: 010584612</p>
+<p>DUNS: 141057898</p>
+<p>Diagnostics for the Real World is a company incorporated and registered under the laws of the State of Delaware (USA).</p>