+++ views/settings/network.php
@@ -1,7 +1,7 @@
 <?php // General Settings
 
 // Hospital Info Settings Subset
-$networkInfoKeys = ['selected_protocol', 'cain_server_ip', 'cain_server_port', 'hl7_server_ip', 'hl7_server_port', 'hl7_server_dest', 'patient_id'];
+$networkInfoKeys = ['selected_protocol', 'cain_server_ip', 'cain_server_port', 'hl7_server_ip', 'hl7_server_port', 'hl7_server_dest', 'patient_id', 'test_mode', 'app_mode'];
 $networkInfo = array_intersect_key($settings, array_flip($networkInfoKeys));
 
 ?>
@@ -40,6 +40,28 @@ $networkInfo = array_intersect_key($settings, array_flip($networkInfoKeys));
                 <input disabled required class="cursor-not-allowed" id="cainPort" type="number" name="cainPort" value="<?= $_SERVER['SERVER_PORT'];?>">
             </div>
         </div>
+    </div>
+    <div class="form-fields">
+        <label for="testMode" class="field !flex-row toggle-field !px-6 py-2 rounded-full bg-white shadow-md">
+            <div class="flex flex-col w-full">
+                <div class="shrink">Test Mode</div>
+                <div class="description !text-xs text-grey mr-4">Dummy data and processes for testing.</div>
+            </div>
+            <div class="checkbox-wrapper">
+                <input class="tgl" name="testMode" id="testMode" type="checkbox" <?= $networkInfo['test_mode'] ? "checked" : "";?>>
+                <label class="toggle" data-tg-off="DISABLED" data-tg-on="ENABLED" for="testMode"><span></span></label>
+            </div>
+        </label>
+        <label for="appMode" class="field !flex-row toggle-field !px-6 py-2 rounded-full bg-white shadow-md">
+            <div class="flex flex-col w-full">
+                <div class="shrink">App Mode</div>
+                <div class="description !text-xs text-grey mr-4">Turn LIMS internal application on/off.</div>
+            </div>
+            <div class="checkbox-wrapper">
+                <input class="tgl" name="appMode" id="appMode" type="checkbox" <?= $networkInfo['app_mode'] ? "checked" : "";?>>
+                <label class="toggle" data-tg-off="DISABLED" data-tg-on="ENABLED" for="appMode"><span></span></label>
+            </div>
+        </label>
     </div>
     <h3 class="text-dark mt-4 w-full text-center rounded-xl px-4 py-2 bg-blue-200/75 shadow-lg">LIMS Settings</h3>
     <div id="hl7Options" class="form-fields <?= $networkInfo['selected_protocol'] == "HL7" ? "active" : "";?>">